save changes

+++ src/com/company/Controller/PlayerDB.java
@@ -47,7 +47,7 @@ public class PlayerDB {
 
         while (resultSet.next()) {
             players.add(new Player(resultSet.getInt("id"), resultSet.getString("username"),
-                    resultSet.getInt("score")));
+                    resultSet.getFloat("score")));
         }
 
         closeConnection();
@@ -59,7 +59,7 @@ public class PlayerDB {
     public static int createUser(Player player) throws SQLException {
 
         makeConnection();
-        statement.execute(String.format("INSERT INTO players (username , score ) value ( '%s' , %d )",
+        statement.execute(String.format("INSERT INTO players (username , score ) value ( '%s' , %f )",
                         player.getUsername() , player.getScore() ),
                 Statement.RETURN_GENERATED_KEYS);
 
@@ -76,7 +76,7 @@ public class PlayerDB {
     public static void editUser(Player player) throws SQLException { //Update in sql
         makeConnection();
 
-        statement.execute(String.format(" UPDATE players SET username = '%s' , score = %d WHERE id = %d ", player.getUsername(), player.getScore(), player.getId()));
+        statement.execute(String.format(" UPDATE players SET username = '%s' , score = %f WHERE id = %d ", player.getUsername(), player.getScore(), player.getId()));
 
         closeConnection();
